Move collections and resoucres in their own package
