+++ tool/tctl/common/resource/collections/authority.go
@@ -1,0 +1,51 @@
+package collections
+
+import (
+	"fmt"
+	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/lib/asciitable"
+	"github.com/gravitational/teleport/lib/sshutils"
+	"github.com/gravitational/trace"
+	"io"
+)
+
+func NewAuthorityCollection(cas []types.CertAuthority) ResourceCollection {
+	return &authorityCollection{cas: cas}
+}
+
+type authorityCollection struct {
+	cas []types.CertAuthority
+}
+
+func (a *authorityCollection) Resources() (r []types.Resource) {
+	for _, resource := range a.cas {
+		r = append(r, resource)
+	}
+	return r
+}
+
+func (a *authorityCollection) WriteText(w io.Writer, verbose bool) error {
+	t := asciitable.MakeTable([]string{"Cluster Name", "CA Type", "Fingerprint", "Role Map"})
+	for _, a := range a.cas {
+		for _, key := range a.GetTrustedSSHKeyPairs() {
+			fingerprint, err := sshutils.AuthorizedKeyFingerprint(key.PublicKey)
+			if err != nil {
+				fingerprint = fmt.Sprintf("<bad key: %v>", err)
+			}
+			var roles string
+			if a.GetType() == types.HostCA {
+				roles = "N/A"
+			} else {
+				roles = fmt.Sprintf("%v", a.CombinedMapping())
+			}
+			t.AddRow([]string{
+				a.GetClusterName(),
+				string(a.GetType()),
+				fingerprint,
+				roles,
+			})
+		}
+	}
+	_, err := t.AsBuffer().WriteTo(w)
+	return trace.Wrap(err)
+}
