+++ tool/tctl/common/resource/collections/externalauditstorage.go
@@ -1,0 +1,46 @@
+package collections
+
+import (
+	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/api/types/externalauditstorage"
+	"github.com/gravitational/teleport/lib/asciitable"
+	"github.com/gravitational/trace"
+	"io"
+)
+
+type externalAuditStorageCollection struct {
+	externalAuditStorages []*externalauditstorage.ExternalAuditStorage
+}
+
+func NewExternalAuditStorageCollection(externalAuditStorages []*externalauditstorage.ExternalAuditStorage) ResourceCollection {
+	return &externalAuditStorageCollection{externalAuditStorages: externalAuditStorages}
+}
+
+func (c *externalAuditStorageCollection) Resources() (r []types.Resource) {
+	for _, a := range c.externalAuditStorages {
+		r = append(r, a)
+	}
+	return r
+}
+
+func (c *externalAuditStorageCollection) WriteText(w io.Writer, verbose bool) error {
+	var rows [][]string
+	for _, a := range c.externalAuditStorages {
+		rows = append(rows, []string{
+			a.GetName(),
+			a.Spec.IntegrationName,
+			a.Spec.PolicyName,
+			a.Spec.Region,
+			a.Spec.SessionRecordingsURI,
+			a.Spec.AuditEventsLongTermURI,
+			a.Spec.AthenaResultsURI,
+			a.Spec.AthenaWorkgroup,
+			a.Spec.GlueDatabase,
+			a.Spec.GlueTable,
+		})
+	}
+	headers := []string{"Name", "IntegrationName", "PolicyName", "Region", "SessionRecordingsURI", "AuditEventsLongTermURI", "AthenaResultsURI", "AthenaWorkgroup", "GlueDatabase", "GlueTable"}
+	t := asciitable.MakeTable(headers, rows...)
+	_, err := t.AsBuffer().WriteTo(w)
+	return trace.Wrap(err)
+}
