+++ tool/tctl/common/resource/collections/lock.go
@@ -1,0 +1,38 @@
+package collections
+
+import (
+	"github.com/gravitational/teleport/api/types"
+	apiutils "github.com/gravitational/teleport/api/utils"
+	"github.com/gravitational/teleport/lib/asciitable"
+	"github.com/gravitational/trace"
+	"io"
+)
+
+type lockCollection struct {
+	locks []types.Lock
+}
+
+func NewLockCollection(locks []types.Lock) ResourceCollection {
+	return &lockCollection{locks: locks}
+}
+
+func (c *lockCollection) Resources() (r []types.Resource) {
+	for _, resource := range c.locks {
+		r = append(r, resource)
+	}
+	return r
+}
+
+func (c *lockCollection) WriteText(w io.Writer, verbose bool) error {
+	t := asciitable.MakeTable([]string{"ID", "Target", "Message", "Expires"})
+	for _, lock := range c.locks {
+		target := lock.Target()
+		expires := "never"
+		if lock.LockExpiry() != nil {
+			expires = apiutils.HumanTimeFormat(*lock.LockExpiry())
+		}
+		t.AddRow([]string{lock.GetName(), target.String(), lock.Message(), expires})
+	}
+	_, err := t.AsBuffer().WriteTo(w)
+	return trace.Wrap(err)
+}
