+++ tool/tctl/common/resource/collections/trustedcluster.go
@@ -1,0 +1,77 @@
+package collections
+
+import (
+	"fmt"
+	"github.com/gravitational/teleport/api/types"
+	apiutils "github.com/gravitational/teleport/api/utils"
+	"github.com/gravitational/teleport/lib/asciitable"
+	"github.com/gravitational/trace"
+	"io"
+	"strconv"
+	"time"
+)
+
+type trustedClusterCollection struct {
+	trustedClusters []types.TrustedCluster
+}
+
+func NewTrustedClusterCollection(trustedClusters []types.TrustedCluster) ResourceCollection {
+	return &trustedClusterCollection{trustedClusters: trustedClusters}
+}
+
+func (c *trustedClusterCollection) Resources() (r []types.Resource) {
+	for _, resource := range c.trustedClusters {
+		r = append(r, resource)
+	}
+	return r
+}
+
+func (c *trustedClusterCollection) WriteText(w io.Writer, verbose bool) error {
+	t := asciitable.MakeTable([]string{
+		"Name", "Enabled", "Token", "Proxy Address", "Reverse Tunnel Address", "Role Map",
+	})
+	for _, tc := range c.trustedClusters {
+		t.AddRow([]string{
+			tc.GetName(),
+			strconv.FormatBool(tc.GetEnabled()),
+			tc.GetToken(),
+			tc.GetProxyAddress(),
+			tc.GetReverseTunnelAddress(),
+			fmt.Sprintf("%v", tc.CombinedMapping()),
+		})
+	}
+	_, err := t.AsBuffer().WriteTo(w)
+	return trace.Wrap(err)
+}
+
+type remoteClusterCollection struct {
+	remoteClusters []types.RemoteCluster
+}
+
+func NewRemoteClusterCollection(remoteClusters []types.RemoteCluster) ResourceCollection {
+	return &remoteClusterCollection{remoteClusters: remoteClusters}
+}
+
+func (c *remoteClusterCollection) Resources() (r []types.Resource) {
+	for _, resource := range c.remoteClusters {
+		r = append(r, resource)
+	}
+	return r
+}
+
+func (c *remoteClusterCollection) WriteText(w io.Writer, verbose bool) error {
+	t := asciitable.MakeTable([]string{"Name", "Status", "Last Heartbeat"})
+	for _, cluster := range c.remoteClusters {
+		lastHeartbeat := cluster.GetLastHeartbeat()
+		t.AddRow([]string{cluster.GetName(), cluster.GetConnectionStatus(), formatLastHeartbeat(lastHeartbeat)})
+	}
+	_, err := t.AsBuffer().WriteTo(w)
+	return trace.Wrap(err)
+}
+
+func formatLastHeartbeat(t time.Time) string {
+	if t.IsZero() {
+		return "not available"
+	}
+	return apiutils.HumanTimeFormat(t)
+}
