+++ tool/tctl/common/resource/resource_health_check_config_test.go
@@ -16,12 +16,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package common
+package resource
 
 import (
 	"bytes"
 	"context"
 	"fmt"
+	"github.com/gravitational/teleport/tool/tctl/common/resource/collections"
 	"os"
 	"path/filepath"
 	"testing"
@@ -126,10 +127,8 @@ func testEditHealthCheckConfig(t *testing.T, clt *authclient.Client) {
 			return trace.Wrap(err, "opening file to edit")
 		}
 		expected.GetMetadata().Revision = created.GetMetadata().GetRevision()
-		collection := &healthCheckConfigCollection{
-			items: []*healthcheckconfigv1.HealthCheckConfig{expected},
-		}
-		return trace.NewAggregate(writeYAML(collection, f), f.Close())
+		collection := collections.NewHealthCheckConfigCollection([]*healthcheckconfigv1.HealthCheckConfig{expected})
+		return trace.NewAggregate(collections.WriteYAML(collection, f), f.Close())
 	}
 
 	// Edit the AutoUpdateConfig resource.
